W-12049491: Refactor tests part 3
