+++ src/test/java/org/mule/extension/ConditionsExclusiveFunctionInvalidTestCase.java
@@ -4,24 +4,24 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.general;
+package org.mule.extension;
 
+import static org.mule.extension.TestVariables.SCHEMA_CONDITIONS_DRAFT201909;
+import static org.mule.extension.TestVariables.SCHEMA_CONDITIONS_DRAFT202012;
+import static org.mule.extension.TestVariables.SCHEMA_CONDITIONS_DRAFT7;
+import static org.hamcrest.CoreMatchers.containsString;
+import static org.junit.Assert.assertThat;
+import static org.junit.rules.ExpectedException.none;
+
+import org.mule.module.json.api.JsonSchemaDereferencingMode;
 import org.hamcrest.BaseMatcher;
 import org.hamcrest.Description;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
-import org.mule.extension.AbstractSchemaValidationTestCase;
-import org.mule.module.json.api.JsonSchemaDereferencingMode;
 
-import static org.hamcrest.CoreMatchers.containsString;
-import static org.junit.Assert.assertThat;
-import static org.junit.rules.ExpectedException.none;
-import static org.mule.extension.TestVariables.*;
+public class ConditionsExclusiveFunctionInvalidTestCase extends AbstractSchemaValidationTestCase {
 
-public class BadObjectValidationTestCase extends AbstractSchemaValidationTestCase {
-
-  private static final String VALIDATOR_FAIL_ON_TRAILING_TOKENS = "jsonSchemaValidator.FailOnTrailingTokens";
 
   private String json;
 
@@ -35,38 +35,22 @@ public class BadObjectValidationTestCase extends AbstractSchemaValidationTestCas
 
   @Override
   protected void doSetUp() throws Exception {
-    json = doGetResource("inputs/bad-object.json");
-    System.setProperty(VALIDATOR_FAIL_ON_TRAILING_TOKENS, "true");
-  }
-
-  @Override
-  protected void doTearDown() {
-    System.clearProperty(VALIDATOR_FAIL_ON_TRAILING_TOKENS);
-  }
-
-  @Test
-  public void Draft34validate() throws Exception {
-    runTestWithSchemaAndValidate(SCHEMA_REQUIRED_OBJECT_ARRAY_DRAFT34);
-  }
-
-  @Test
-  public void Draft6validate() throws Exception {
-    runTestWithSchemaAndValidate(SCHEMA_REQUIRED_OBJECT_ARRAY_DRAFT6);
+    json = doGetResource("inputs/drarft-07-orGreater-exclusive-function-conditions-INVALID.json");
   }
 
   @Test
   public void Draft7validate() throws Exception {
-    runTestWithSchemaAndValidate(SCHEMA_REQUIRED_OBJECT_ARRAY_DRAFT7);
+    runTestWithSchemaAndValidate(SCHEMA_CONDITIONS_DRAFT7);
   }
 
   @Test
   public void Draft201909validate() throws Exception {
-    runTestWithSchemaAndValidate(SCHEMA_REQUIRED_OBJECT_ARRAY_DRAFT201909);
+    runTestWithSchemaAndValidate(SCHEMA_CONDITIONS_DRAFT201909);
   }
 
   @Test
-  public void Draft2020129validate() throws Exception {
-    runTestWithSchemaAndValidate(SCHEMA_REQUIRED_OBJECT_ARRAY_DRAFT202012);
+  public void Draft202012validate() throws Exception {
+    runTestWithSchemaAndValidate(SCHEMA_CONDITIONS_DRAFT202012);
   }
 
   private void runTestWithSchemaAndValidate(String schema) throws Exception {
@@ -77,7 +61,7 @@ public class BadObjectValidationTestCase extends AbstractSchemaValidationTestCas
       public boolean matches(Object item) {
         Exception e = (Exception) item;
         String report = e.getMessage();
-        assertThat(report, containsString("Trailing token (of type START_OBJECT) found after value"));
+        assertThat(report, containsString("$.bar: is missing but it is required"));
 
         return true;
       }
