+++ src/test/java/org/mule/extension/TestVariables.java
@@ -9,16 +9,17 @@ package org.mule.extension;
 public class TestVariables {
 
   //Schemas
-  //Draft 4 & 3
-  public static final String SCHEMA_FSTAB_JSON_DRAFT34 = "/Draft34/schemas/fstab.json";
-  public static final String SCHEMA_FSTAB_DRAFTV3 = "/Draft34/schemas/fstab-draftv3.json";
-  public static final String SCHEMA_FSTAB_INLINE_DRAFT34 = "/Draft34/schemas/fstab-inline.json";
-  public static final String SCHEMA_FSTAB_REFERRING_DRAFT34 = "/Draft34/schemas/fstab-referring.json";
-  public static final String SCHEMA_FSTAB_ARBITRARY_PRECISION_KEYS_DRAFT34 =
-      "Draft34/schemas/fstab-arbitrary-precision-keys.json";
-  public static final String SCHEMA_REQUIRED_OBJECT_ARRAY_DRAFT34 = "/Draft34/schemas/object-array-schema.json";
+  //Draft 3
+  public static final String SCHEMA_FSTAB_DRAFTV3 = "/Draft4/schemas/fstab-draftv3.json";
 
-  //Schemas
+  //Draft 4
+  public static final String SCHEMA_FSTAB_JSON_DRAFT4 = "/Draft4/schemas/fstab.json";
+  public static final String SCHEMA_FSTAB_INLINE_DRAFT4 = "/Draft4/schemas/fstab-inline.json";
+  public static final String SCHEMA_FSTAB_REFERRING_DRAFT4 = "/Draft4/schemas/fstab-referring.json";
+  public static final String SCHEMA_FSTAB_ARBITRARY_PRECISION_KEYS_DRAFT4 =
+      "Draft4/schemas/fstab-arbitrary-precision-keys.json";
+  public static final String SCHEMA_REQUIRED_OBJECT_ARRAY_DRAFT4 = "/Draft4/schemas/object-array-schema.json";
+
   //Draft 6
   public static final String SCHEMA_FSTAB_JSON_DRAFT6 = "/Draft6/schemas/fstab.json";
   public static final String SCHEMA_FSTAB_INLINE_DRAFT6 = "/Draft6/schemas/fstab-inline.json";
@@ -27,7 +28,6 @@ public class TestVariables {
       "Draft6/schemas/fstab-arbitrary-precision-keys.json";
   public static final String SCHEMA_REQUIRED_OBJECT_ARRAY_DRAFT6 = "/Draft6/schemas/object-array-schema.json";
 
-  //Schemas
   //Draft 7
   public static final String SCHEMA_FSTAB_JSON_DRAFT7 = "/Draft7/schemas/fstab.json";
   public static final String SCHEMA_FSTAB_INLINE_DRAFT7 = "/Draft7/schemas/fstab-inline.json";
@@ -38,7 +38,6 @@ public class TestVariables {
 
   public static final String SCHEMA_CONDITIONS_DRAFT7 = "/Draft7/schemas/conditions-schema.json";
 
-  //Schemas
   //Draft 201909
   public static final String SCHEMA_FSTAB_JSON_DRAFT201909 = "/Draft201909/schemas/fstab.json";
   public static final String SCHEMA_FSTAB_INLINE_DRAFT201909 = "/Draft201909/schemas/fstab-inline.json";
@@ -50,7 +49,6 @@ public class TestVariables {
 
   public static final String SCHEMA_DEPENDENT_DRAFT2019009 = "/Draft201909/schemas/dependent-schema.json";
 
-  //Schemas
   //Draft 202012
   public static final String SCHEMA_FSTAB_JSON_DRAFT202012 = "/Draft202012/schemas/fstab.json";
   public static final String SCHEMA_FSTAB_INLINE_DRAFT202012 = "/Draft202012/schemas/fstab-inline.json";
@@ -60,7 +58,4 @@ public class TestVariables {
   public static final String SCHEMA_REQUIRED_OBJECT_ARRAY_DRAFT202012 = "/Draft202012/schemas/object-array-schema.json";
   public static final String SCHEMA_CONDITIONS_DRAFT202012 = "/Draft202012/schemas/conditions-schema.json";
   public static final String SCHEMA_DEPENDENT_DRAFT202012 = "/Draft202012/schemas/dependent-schema.json";
-
-
-
 }
