+++ src/test/java/org/mule/extension/ValidateArbitraryPrecisionTestCase.java
@@ -4,25 +4,24 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.extension.general;
+package org.mule.extension;
 
+import static org.mule.extension.TestVariables.SCHEMA_FSTAB_ARBITRARY_PRECISION_KEYS_DRAFT201909;
+import static org.mule.extension.TestVariables.SCHEMA_FSTAB_ARBITRARY_PRECISION_KEYS_DRAFT202012;
+import static org.mule.extension.TestVariables.SCHEMA_FSTAB_ARBITRARY_PRECISION_KEYS_DRAFT4;
+import static org.mule.extension.TestVariables.SCHEMA_FSTAB_ARBITRARY_PRECISION_KEYS_DRAFT6;
+import static org.mule.extension.TestVariables.SCHEMA_FSTAB_ARBITRARY_PRECISION_KEYS_DRAFT7;
+import static java.util.Arrays.asList;
+import static org.hamcrest.core.StringContains.containsString;
+import static org.junit.rules.ExpectedException.none;
+
+import org.mule.test.runner.RunnerDelegateTo;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.internal.matchers.ThrowableMessageMatcher;
 import org.junit.rules.ExpectedException;
 import org.junit.runners.Parameterized;
-import org.mule.extension.AbstractSchemaValidationTestCase;
-import org.mule.module.json.api.JsonSchemaDereferencingMode;
-import org.mule.runtime.core.api.event.CoreEvent;
-import org.mule.test.runner.RunnerDelegateTo;
-
 import java.util.Collection;
-
-import static java.util.Arrays.asList;
-import static org.hamcrest.core.StringContains.containsString;
-import static org.junit.Assert.assertEquals;
-import static org.junit.rules.ExpectedException.none;
-import static org.mule.extension.TestVariables.*;
 
 @RunnerDelegateTo(Parameterized.class)
 public class ValidateArbitraryPrecisionTestCase extends AbstractSchemaValidationTestCase {
@@ -54,7 +53,7 @@ public class ValidateArbitraryPrecisionTestCase extends AbstractSchemaValidation
 
   @Test
   public void Draft4AllowArbitraryPrecision() throws Exception {
-    runTestWithSchemaAndValidate(SCHEMA_FSTAB_ARBITRARY_PRECISION_KEYS_DRAFT34);
+    runTestWithSchemaAndValidate(SCHEMA_FSTAB_ARBITRARY_PRECISION_KEYS_DRAFT4);
   }
 
   @Test
