--- src/test/java/org/mule/extension/general/ConditionsExclusiveFunctionInvalidTestCase.java
@@ -1,79 +1,0 @@
-/*
- * Copyright (c) MuleSoft, Inc.  All rights reserved.  http://www.mulesoft.com
- * The software in this package is published under the terms of the CPAL v1.0
- * license, a copy of which has been included with this distribution in the
- * LICENSE.txt file.
- */
-package org.mule.extension.general;
-
-import org.hamcrest.BaseMatcher;
-import org.hamcrest.Description;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.rules.ExpectedException;
-import org.mule.extension.AbstractSchemaValidationTestCase;
-import org.mule.module.json.api.JsonSchemaDereferencingMode;
-
-import static org.hamcrest.CoreMatchers.containsString;
-import static org.junit.Assert.assertThat;
-import static org.junit.rules.ExpectedException.none;
-import static org.mule.extension.TestVariables.*;
-
-public class ConditionsExclusiveFunctionInvalidTestCase extends AbstractSchemaValidationTestCase {
-
-
-  private String json;
-
-  @Rule
-  public ExpectedException expectedException = none();
-
-  @Override
-  protected String getConfigFile() {
-    return "config/schema-validation-config.xml";
-  }
-
-  @Override
-  protected void doSetUp() throws Exception {
-    json = doGetResource("inputs/drarft-07-orGreater-exclusive-function-conditions-INVALID.json");
-  }
-
-  @Test
-  public void Draft7validate() throws Exception {
-    runTestWithSchemaAndValidate(SCHEMA_CONDITIONS_DRAFT7);
-  }
-
-  @Test
-  public void Draft201909validate() throws Exception {
-    runTestWithSchemaAndValidate(SCHEMA_CONDITIONS_DRAFT201909);
-  }
-
-  @Test
-  public void Draft202012validate() throws Exception {
-    runTestWithSchemaAndValidate(SCHEMA_CONDITIONS_DRAFT202012);
-  }
-
-  private void runTestWithSchemaAndValidate(String schema) throws Exception {
-
-    expectedException.expectCause(new BaseMatcher<Throwable>() {
-
-      @Override
-      public boolean matches(Object item) {
-        Exception e = (Exception) item;
-        String report = e.getMessage();
-        assertThat(report, containsString("$.bar: is missing but it is required"));
-
-        return true;
-      }
-
-      @Override
-      public void describeTo(Description description) {
-        description.appendText("Error report did not match");
-      }
-    });
-
-    flowRunner("validate")
-        .withVariable("schema", schema)
-        .withVariable("dereferencing", JsonSchemaDereferencingMode.CANONICAL)
-        .withPayload(json).run();
-  }
-}
