+++ src/test/resources/Draft202012/schemas/schema-content-references.json
@@ -1,0 +1,18 @@
+{
+  "$schema": "http://json-schema.org/draft/2020-12/schema#",
+  "title": "/etc/fstab",
+  "description": "JSON representation of /etc/fstab",
+  "type": "object",
+  "properties": {
+    "swap": {
+      "$ref": "fstab.json#/definitions/mntent"
+    }
+  },
+  "patternProperties": {
+    "^/([^/]+(/[^/]+)*)?$": {
+      "$ref": "fstab.json#/definitions/mntent"
+    }
+  },
+  "required": [ "/", "swap" ],
+  "additionalProperties": false
+}
