+++ src/test/resources/Draft4/schemas/fstab-arbitrary-precision-keys.json
