+++ src/test/resources/Draft4/schemas/fstab.json
