+++ src/test/resources/Draft4/schemas/object-array-schema.json
