+++ src/test/resources/Draft4/schemas/schema-default.json
